+++ sandbox/libs/dataformat-native/src/main/java/org/opensearch/nativebridge/spi/NativeMemoryFetcher.java
@@ -10,6 +10,7 @@ package org.opensearch.nativebridge.spi;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+import org.opensearch.plugin.stats.NativeMemoryStats;
 
 import java.lang.foreign.FunctionDescriptor;
 import java.lang.foreign.Linker;
