+++ sandbox/libs/dataformat-native/src/test/java/org/opensearch/nativebridge/spi/NativeMemoryStatsSerializationTests.java
@@ -10,6 +10,7 @@ package org.opensearch.nativebridge.spi;
 
 import org.opensearch.common.io.stream.BytesStreamOutput;
 import org.opensearch.core.common.io.stream.StreamInput;
+import org.opensearch.plugin.stats.NativeMemoryStats;
 import org.opensearch.test.OpenSearchTestCase;
 
 import java.io.IOException;
