+++ sandbox/modules/native-bridge/src/main/java/org/opensearch/nativebridge/NativeBridgeModule.java
@@ -22,8 +22,8 @@ import org.opensearch.nativebridge.spi.NativeAllocatorConfig;
 import org.opensearch.nativebridge.spi.NativeHeapProfiler;
 import org.opensearch.nativebridge.spi.NativeLibraryLoader;
 import org.opensearch.nativebridge.spi.NativeMemoryFetcher;
-import org.opensearch.nativebridge.spi.NativeMemoryStats;
-import org.opensearch.nativebridge.spi.NativeStatsProvider;
+import org.opensearch.plugin.stats.NativeMemoryStats;
+import org.opensearch.plugin.stats.NativeStatsProvider;
 import org.opensearch.plugins.Plugin;
 import org.opensearch.repositories.RepositoriesService;
 import org.opensearch.script.ScriptService;
